Implements: -  new scritpt chapter 10 -  change role about next test

+++ src/main/java/com/speakto/english/model/FrameEntity.java
@@ -19,5 +19,5 @@ public class FrameEntity {
     @Column(name = "next_test")
     private LocalDateTime nextTest;
     @Column(name = "num_right_answer")
-    private Integer minutesToNextTest;
+    private Long minutesToNextTest;
 }
